Update to add comments to commit
For this first program commit, I began coding by first laying down all the user input strings and their variables that will be used for the calculations. I also added in most of the print statements the program will show the user, with the exception of the print statements that return the calculations and a couple of others I plan to add at the end of the program to remind the user to re-check their blood sugars again a little later. I am leaving the second half of the original program in as placeholders for some of the loops and other programming concepts I might use when building this program.