CalcBolus update: Commit #2
Updated the CalcBolus program with the first additions to the program. These additions include initializing variables for the first round of while loops and decision trees as well as general code cleanup and grammar fixes. I also removed and edited some of the cloned project's code to fit in with my program, but left some empty while statements and operations that I'll be using later.