CalcBolus Commit #3
For this most recent commit I worked on building the program more by adding if/elif/else statements and while to the necessary user input strings for two reasons. The first being to add a pathway for the input to be placed into the initialized variables and the second being to make sure the input the user is giving is valid. I added the remaining variables that needed to be initizalized in the second half of the program, and finished the first round of decision trees for all of the yes/no questions without yet adding the code for calculating any of the numbers, which will begin with my next commit. Lastly, I edited some of the wording ad added some "\n" spacing for legibility.