CalcBolus Commit #4
For the fourth commit, I both added to and edited existing code. First, I made minor changes to variable names. I added in the remianing user input strings and variables like target blood glucose and insulin to carb ratio. I also created the first round of calculations as per the project roadmap item 8 that calculates the base bolus units without accounting for phyiscal factors using while loops and if/else statements. I also tried adding bold text to the program but was unsuccessful and it looks like I will have to import a color or font module to do so, whichI will consider doing as I add the finishing touches to the program.