Update projectroadmap.md
Removed an item that I deemed unnecessary for the program's progress checkpoints and added to item 11, stating that I would look into adding text colors or bolding to the program once the main code is complete and correctly running. I also marked off item 8, which was completed in my fourth commit of the calcbolus.py program.