Commit #5
This commit adds the first part of the if/elif/else statements for the formulas that account for physical states. I also changed the base formula that does not account for those factors to merge all of these calculations into one while statement to produce one final variable, named bolus. The first part of these statements is for the base formula that is used if no physical factors are present, and the second is for hypoglycemia causing factors that use the variable "lowfact", as well as the "highfact" variable when necessary. I plan to mirror this setup for the hyperglycemia causing factors in the next commit and possibly do some code cleanup as well.