commit #6
For the sixth commit, the program was finished, and you could call it the "rought draft" in that it has all the elements it needs but is not what the final submission will be and will still need editing and revision. I added the last of the if/elif/else statements that trigger with the condition that hyperglycemia-causing events are documented by the user. I also added the last else statement to close out the whole decision tree for the calculations. The program has been tested to make sure the operations are syntactically correct and follow mathematic order rules.